Supprime la persistance des informations générales
